Remove in-house upgrade feature & restore dump to 3.1.3000 version
Upgrading to versions that are still under development is
brittle and unlikely to work for a long time thus the
upgrade feature (to 4.0.0) is removed.

The dump files of tpcds30tb dataset were wrongly upgraded to
version 4.0.0 by commit (f8c444bbf935c45bb7a3492393c8289876ab0d13);
this commit restores the respective dumps to version 3.1.3000.